HCK-4090 migrate to new build system (#10)
This plugin still has app.require('aws-sdk'). The reason is this plugin
requires an old version of aws-sdk which is large, so if we install it
into the plugin - the release build with be around 10 MB instead of 3MB.

Until we migrate to aws-sdk v3 we will use app.require('aws-sdk').